refactor: move extractApiError helper to shared location and enhance error handling
- Moved extractApiError function from invite-your-team-service to helpers/extract-api-error.js for reusability
- Enhanced extract-api-error helper to handle nested error structures with detail and source pointer fields
- Added support for errors array format with field extraction from source pointer
- Removed duplicate extractApiError implementation from invite-your-team-service
- Removed custom field-specific error handling logic
